requirements.txt
amqp==2.2.2
Babel==2.5.3
billiard==3.5.0.3
celery==4.1.0
Django==1.11.11
djangorestframework==3.7.7
flower==0.9.2
kombu==4.1.0
#psycopg2==2.7.4
pytz==2018.3
requests==2.18.4
tornado==5.0.1
vine==1.1.4
